Project clean-up

+++ CanvasTest/src/main/java/com/google/code/broccolis/xydroid/activities/LevelActivity.java
@@ -9,7 +9,6 @@ import android.view.View;
 import android.widget.AdapterView;
 import android.widget.ArrayAdapter;
 import android.widget.Button;
-import android.widget.FrameLayout;
 import android.widget.ListView;
 import android.widget.RelativeLayout;
 
@@ -84,101 +83,101 @@ public class LevelActivity extends Activity
 
     private void setComponentSizes()
     {
-        Button button = (Button)findViewById(R.id.button11);
-        button.setScaleX((float) SCREEN_WIDTH / (NEXUS_WIDTH));
-        button.setScaleY((float) SCREEN_HEIGHT / (NEXUS_HEIGHT));
+        Button button = (Button) findViewById(R.id.button11);
+        button.setScaleX(SCREEN_WIDTH / (NEXUS_WIDTH));
+        button.setScaleY(SCREEN_HEIGHT / (NEXUS_HEIGHT));
 
-        button = (Button)findViewById(R.id.button12);
-        button.setScaleX((float) SCREEN_WIDTH / (NEXUS_WIDTH));
-        button.setScaleY((float) SCREEN_HEIGHT / (NEXUS_HEIGHT));
+        button = (Button) findViewById(R.id.button12);
+        button.setScaleX(SCREEN_WIDTH / (NEXUS_WIDTH));
+        button.setScaleY(SCREEN_HEIGHT / (NEXUS_HEIGHT));
 
-        button = (Button)findViewById(R.id.button13);
-        button.setScaleX((float) SCREEN_WIDTH / (NEXUS_WIDTH));
-        button.setScaleY((float) SCREEN_HEIGHT / (NEXUS_HEIGHT));
+        button = (Button) findViewById(R.id.button13);
+        button.setScaleX(SCREEN_WIDTH / (NEXUS_WIDTH));
+        button.setScaleY(SCREEN_HEIGHT / (NEXUS_HEIGHT));
 
-        button = (Button)findViewById(R.id.button21);
-        button.setScaleX((float) SCREEN_WIDTH / (NEXUS_WIDTH));
-        button.setScaleY((float) SCREEN_HEIGHT / (NEXUS_HEIGHT));
+        button = (Button) findViewById(R.id.button21);
+        button.setScaleX(SCREEN_WIDTH / (NEXUS_WIDTH));
+        button.setScaleY(SCREEN_HEIGHT / (NEXUS_HEIGHT));
 
-        button = (Button)findViewById(R.id.button22);
-        button.setScaleX((float) SCREEN_WIDTH / (NEXUS_WIDTH));
-        button.setScaleY((float) SCREEN_HEIGHT / (NEXUS_HEIGHT));
+        button = (Button) findViewById(R.id.button22);
+        button.setScaleX(SCREEN_WIDTH / (NEXUS_WIDTH));
+        button.setScaleY(SCREEN_HEIGHT / (NEXUS_HEIGHT));
 
-        button = (Button)findViewById(R.id.button23);
-        button.setScaleX((float) SCREEN_WIDTH / (NEXUS_WIDTH));
-        button.setScaleY((float) SCREEN_HEIGHT / (NEXUS_HEIGHT));
+        button = (Button) findViewById(R.id.button23);
+        button.setScaleX(SCREEN_WIDTH / (NEXUS_WIDTH));
+        button.setScaleY(SCREEN_HEIGHT / (NEXUS_HEIGHT));
 
-        button = (Button)findViewById(R.id.button31);
-        button.setScaleX((float) SCREEN_WIDTH / (NEXUS_WIDTH));
-        button.setScaleY((float) SCREEN_HEIGHT / (NEXUS_HEIGHT));
+        button = (Button) findViewById(R.id.button31);
+        button.setScaleX(SCREEN_WIDTH / (NEXUS_WIDTH));
+        button.setScaleY(SCREEN_HEIGHT / (NEXUS_HEIGHT));
 
-        button = (Button)findViewById(R.id.button32);
-        button.setScaleX((float) SCREEN_WIDTH / (NEXUS_WIDTH));
-        button.setScaleY((float) SCREEN_HEIGHT / (NEXUS_HEIGHT));
+        button = (Button) findViewById(R.id.button32);
+        button.setScaleX(SCREEN_WIDTH / (NEXUS_WIDTH));
+        button.setScaleY(SCREEN_HEIGHT / (NEXUS_HEIGHT));
 
-        button = (Button)findViewById(R.id.button33);
-        button.setScaleX((float) SCREEN_WIDTH / (NEXUS_WIDTH));
-        button.setScaleY((float) SCREEN_HEIGHT / (NEXUS_HEIGHT));
+        button = (Button) findViewById(R.id.button33);
+        button.setScaleX(SCREEN_WIDTH / (NEXUS_WIDTH));
+        button.setScaleY(SCREEN_HEIGHT / (NEXUS_HEIGHT));
 
-        button = (Button)findViewById(R.id.button41);
-        button.setScaleX((float) SCREEN_WIDTH / (NEXUS_WIDTH));
-        button.setScaleY((float) SCREEN_HEIGHT / (NEXUS_HEIGHT));
+        button = (Button) findViewById(R.id.button41);
+        button.setScaleX(SCREEN_WIDTH / (NEXUS_WIDTH));
+        button.setScaleY(SCREEN_HEIGHT / (NEXUS_HEIGHT));
 
-        button = (Button)findViewById(R.id.button42);
-        button.setScaleX((float) SCREEN_WIDTH / (NEXUS_WIDTH));
-        button.setScaleY((float) SCREEN_HEIGHT / (NEXUS_HEIGHT));
+        button = (Button) findViewById(R.id.button42);
+        button.setScaleX(SCREEN_WIDTH / (NEXUS_WIDTH));
+        button.setScaleY(SCREEN_HEIGHT / (NEXUS_HEIGHT));
 
-        button = (Button)findViewById(R.id.button43);
-        button.setScaleX((float) SCREEN_WIDTH / (NEXUS_WIDTH));
-        button.setScaleY((float) SCREEN_HEIGHT / (NEXUS_HEIGHT));
+        button = (Button) findViewById(R.id.button43);
+        button.setScaleX(SCREEN_WIDTH / (NEXUS_WIDTH));
+        button.setScaleY(SCREEN_HEIGHT / (NEXUS_HEIGHT));
 
-        button = (Button)findViewById(R.id.button61);
-        button.setScaleX((float) SCREEN_WIDTH / (NEXUS_WIDTH));
-        button.setScaleY((float) SCREEN_HEIGHT / (NEXUS_HEIGHT));
+        button = (Button) findViewById(R.id.button61);
+        button.setScaleX(SCREEN_WIDTH / (NEXUS_WIDTH));
+        button.setScaleY(SCREEN_HEIGHT / (NEXUS_HEIGHT));
 
-        button = (Button)findViewById(R.id.button62);
-        button.setScaleX((float) SCREEN_WIDTH / (NEXUS_WIDTH));
-        button.setScaleY((float) SCREEN_HEIGHT / (NEXUS_HEIGHT));
+        button = (Button) findViewById(R.id.button62);
+        button.setScaleX(SCREEN_WIDTH / (NEXUS_WIDTH));
+        button.setScaleY(SCREEN_HEIGHT / (NEXUS_HEIGHT));
 
-        button = (Button)findViewById(R.id.button63);
-        button.setScaleX((float) SCREEN_WIDTH / (NEXUS_WIDTH));
-        button.setScaleY((float) SCREEN_HEIGHT / (NEXUS_HEIGHT));
+        button = (Button) findViewById(R.id.button63);
+        button.setScaleX(SCREEN_WIDTH / (NEXUS_WIDTH));
+        button.setScaleY(SCREEN_HEIGHT / (NEXUS_HEIGHT));
 
-        button = (Button)findViewById(R.id.button71);
-        button.setScaleX((float) SCREEN_WIDTH / (NEXUS_WIDTH));
-        button.setScaleY((float) SCREEN_HEIGHT / (NEXUS_HEIGHT));
+        button = (Button) findViewById(R.id.button71);
+        button.setScaleX(SCREEN_WIDTH / (NEXUS_WIDTH));
+        button.setScaleY(SCREEN_HEIGHT / (NEXUS_HEIGHT));
 
-        button = (Button)findViewById(R.id.button72);
-        button.setScaleX((float) SCREEN_WIDTH / (NEXUS_WIDTH));
-        button.setScaleY((float) SCREEN_HEIGHT / (NEXUS_HEIGHT));
+        button = (Button) findViewById(R.id.button72);
+        button.setScaleX(SCREEN_WIDTH / (NEXUS_WIDTH));
+        button.setScaleY(SCREEN_HEIGHT / (NEXUS_HEIGHT));
 
-        button = (Button)findViewById(R.id.button73);
-        button.setScaleX((float) SCREEN_WIDTH / (NEXUS_WIDTH));
-        button.setScaleY((float) SCREEN_HEIGHT / (NEXUS_HEIGHT));
+        button = (Button) findViewById(R.id.button73);
+        button.setScaleX(SCREEN_WIDTH / (NEXUS_WIDTH));
+        button.setScaleY(SCREEN_HEIGHT / (NEXUS_HEIGHT));
 
-        button = (Button)findViewById(R.id.button81);
-        button.setScaleX((float) SCREEN_WIDTH / (NEXUS_WIDTH));
-        button.setScaleY((float) SCREEN_HEIGHT / (NEXUS_HEIGHT));
+        button = (Button) findViewById(R.id.button81);
+        button.setScaleX(SCREEN_WIDTH / (NEXUS_WIDTH));
+        button.setScaleY(SCREEN_HEIGHT / (NEXUS_HEIGHT));
 
-        button = (Button)findViewById(R.id.button82);
-        button.setScaleX((float) SCREEN_WIDTH / (NEXUS_WIDTH));
-        button.setScaleY((float) SCREEN_HEIGHT / (NEXUS_HEIGHT));
+        button = (Button) findViewById(R.id.button82);
+        button.setScaleX(SCREEN_WIDTH / (NEXUS_WIDTH));
+        button.setScaleY(SCREEN_HEIGHT / (NEXUS_HEIGHT));
 
-        button = (Button)findViewById(R.id.button83);
-        button.setScaleX((float) SCREEN_WIDTH / (NEXUS_WIDTH));
-        button.setScaleY((float) SCREEN_HEIGHT / (NEXUS_HEIGHT));
+        button = (Button) findViewById(R.id.button83);
+        button.setScaleX(SCREEN_WIDTH / (NEXUS_WIDTH));
+        button.setScaleY(SCREEN_HEIGHT / (NEXUS_HEIGHT));
 
-        button = (Button)findViewById(R.id.button91);
-        button.setScaleX((float) SCREEN_WIDTH / (NEXUS_WIDTH));
-        button.setScaleY((float) SCREEN_HEIGHT / (NEXUS_HEIGHT));
+        button = (Button) findViewById(R.id.button91);
+        button.setScaleX(SCREEN_WIDTH / (NEXUS_WIDTH));
+        button.setScaleY(SCREEN_HEIGHT / (NEXUS_HEIGHT));
 
-        button = (Button)findViewById(R.id.button92);
-        button.setScaleX((float) SCREEN_WIDTH / (NEXUS_WIDTH));
-        button.setScaleY((float) SCREEN_HEIGHT / (NEXUS_HEIGHT));
+        button = (Button) findViewById(R.id.button92);
+        button.setScaleX(SCREEN_WIDTH / (NEXUS_WIDTH));
+        button.setScaleY(SCREEN_HEIGHT / (NEXUS_HEIGHT));
 
-        button = (Button)findViewById(R.id.button93);
-        button.setScaleX((float) SCREEN_WIDTH / (NEXUS_WIDTH));
-        button.setScaleY((float) SCREEN_HEIGHT / (NEXUS_HEIGHT));
+        button = (Button) findViewById(R.id.button93);
+        button.setScaleX(SCREEN_WIDTH / (NEXUS_WIDTH));
+        button.setScaleY(SCREEN_HEIGHT / (NEXUS_HEIGHT));
     }
 
     @Override
