+++ CanvasTest/src/main/java/com/google/code/broccolis/xydroid/util/MultipleDeviceSupport.java
@@ -9,17 +9,17 @@ public class MultipleDeviceSupport
     public static final float NEXUS_WIDTH = 1280;
     public static final float NEXUS_HEIGHT = 736;
     private static final String NAME = "MultipleDeviceSupport ";
-    public static int SCREEN_WIDTH;
-    public static int SCREEN_HEIGHT;
+    public static float SCREEN_WIDTH;
+    public static float SCREEN_HEIGHT;
 
     public static float parseXToFloat(int x) throws IllegalArgumentException
     {
-        return ((float) x) / ((float) SCREEN_WIDTH);
+        return ((float) x) / (SCREEN_WIDTH);
     }
 
     public static float parseYToFloat(int y) throws IllegalArgumentException
     {
-        return ((float) y) / ((float) SCREEN_HEIGHT);
+        return ((float) y) / (SCREEN_HEIGHT);
     }
 
     public static int parseXToInt(float x)
